+++ tests/profiling/catalog_sales.preql
