+++ tests/profiling/customer.preql
@@ -1,0 +1,43 @@
+import customer_demographic as demographics;
+
+
+key id int;
+property id.text_id string;
+property id.last_name string;
+property id.first_name string;
+property id.preferred_cust_flag string;
+
+key address_id int;
+property address_id.address_id_string string;
+property address_id.street string;
+property address_id.city string;
+property address_id.state string?;
+property address_id.zip string;
+property address_id.county string;
+
+
+
+datasource customers (
+    C_CUSTOMER_SK: id,
+    C_CUSTOMER_ID: text_id,
+    C_LAST_NAME: last_name,
+    C_FIRST_NAME: first_name,
+    C_CURRENT_ADDR_SK: address_id,
+    C_CURRENT_CDEMO_SK: demographics.id,
+    C_PREFERRED_CUST_FLAG: preferred_cust_flag
+)
+grain (id)
+address memory.customer;
+
+
+datasource customer_address(
+    CA_ADDRESS_SK: address_id,
+    CA_STREET: street,
+    CA_CITY: city,
+    CA_STATE: state,
+    CA_ZIP: zip,
+    CA_COUNTY: county,
+)
+grain (address_id)
+address memory.customer_address;
+
