+++ tests/profiling/customer_demographic.preql
@@ -1,0 +1,25 @@
+
+key id int;
+property id.gender string;
+property id.marital_status string;
+property id.education_status string;
+property id.purchase_estimate string;
+property id.credit_rating string;
+property id.dependent_count int;
+property id.employed_dependent_count int;
+property id.college_dependent_count int; 
+
+
+datasource customer_demographics (
+        CD_DEMO_SK: id,
+        CD_GENDER: gender,
+        CD_MARITAL_STATUS: marital_status,
+        CD_EDUCATION_STATUS: education_status,
+        CD_PURCHASE_ESTIMATE: purchase_estimate,
+        CD_CREDIT_RATING: credit_rating,
+        CD_DEP_COUNT: dependent_count,
+        CD_DEP_EMPLOYED_COUNT: employed_dependent_count,
+        CD_DEP_COLLEGE_COUNT: college_dependent_count,
+)
+grain (id)
+address memory.customer_demographics;
