+++ tests/profiling/promotion.preql
@@ -1,0 +1,17 @@
+key id int;
+property id.string_id string;
+property id.channel_email string;
+property id.channel_catalog string;
+property id.channel_event string;
+property id.cost float;
+
+datasource promotion (
+    P_PROMO_SK: id,
+    P_PROMO_ID: string_id,
+    P_COST: cost,
+    P_CHANNEL_EMAIL: channel_email,
+    P_CHANNEL_EVENT: channel_event,
+    P_CHANNEL_CATALOG: channel_catalog,
+)
+grain (id)
+address memory.promotion;
