+++ tests/profiling/query01.preql
@@ -1,0 +1,30 @@
+import customer as customer;
+import store as store;
+import store_returns as returns;
+
+# query 1
+# Find customers who have returned items more than 20% more often than the average customer returns for a
+# store in a given state for a given year.
+rowset tn_2000 <-select
+    returns.customer.text_id,
+    returns.store.id,
+    returns.store.state,
+    returns.return_date.year,
+    sum(returns.return_amount)-> total_returns
+where
+    returns.return_date.year = 2000
+    and returns.store.state = 'TN';
+
+auto avg_store_returns <- avg(tn_2000.total_returns) by tn_2000.returns.store.id;
+
+select
+    tn_2000.returns.customer.text_id,
+    --tn_2000.total_returns,
+    --avg_store_returns,
+where
+    tn_2000.total_returns > (1.2*avg_store_returns)
+order by 
+    tn_2000.returns.customer.text_id asc
+limit 100;
+
+
