+++ tests/profiling/query03.preql
@@ -1,0 +1,18 @@
+import store_sales as store_sales;
+
+
+#Report the total extended sales price per item brand of a specific manufacturer for all sales in a specific month
+# of the year
+
+SELECT
+    store_sales.date.year,
+    store_sales.item.brand_id,
+    store_sales.item.brand_name,
+    sum(filter store_sales.ext_sales_price where store_sales.date.month_of_year=11 and store_sales.item.manufacturer_id=128)->sum_agg
+WHERE 
+    sum_agg>0
+ORDER BY 
+    store_sales.date.year asc,
+    sum_agg desc,
+    store_sales.item.brand_id asc
+limit 100;
