+++ tests/profiling/query10.preql
@@ -1,0 +1,72 @@
+import customer as customer;
+import store_sales as store_sales;
+import web_sales as web_sales;
+import catalog_sales as catalog_sales;
+
+#Count the customers with the same gender, marital status, education status, purchase estimate, credit rating,
+#dependent count, employed dependent count and college dependent count who live in certain counties and who
+# have purchased from both stores and another sales channel during a three month time period of a given year.
+
+merge catalog_sales.customer.id, web_sales.customer.id, store_sales.customer.id, customer.id;
+
+
+auto relevant_customers <- filter 
+    customer.id 
+where 
+    store_sales.date.year = 2002 and store_sales.date.month_of_year in (1,2,3,4)
+        and ((web_sales.date.year = 2002 and web_sales.date.month_of_year in (1,2,3,4)) or (catalog_sales.date.year = 2002 and catalog_sales.date.month_of_year in (1,2,3,4)))
+    and customer.county in (
+    'Rush County',
+    'Toole County',
+    'Jefferson County',
+    'Dona Ana County',
+    'La Porte County',
+);
+
+with base as select 
+    customer.id,
+    customer.demographics.gender,
+    customer.demographics.marital_status,
+    customer.demographics.education_status,
+    customer.demographics.purchase_estimate,
+    customer.demographics.credit_rating,
+    customer.demographics.dependent_count,
+    customer.demographics.employed_dependent_count,
+    customer.demographics.college_dependent_count,
+WHERE
+customer.id in relevant_customers
+;
+
+SELECT
+    base.customer.demographics.gender,
+    base.customer.demographics.marital_status,
+    base.customer.demographics.education_status,
+    count(base.customer.id
+     ) as cnt1,
+    base.customer.demographics.purchase_estimate,
+        count(base.customer.id
+     ) as cnt2,
+    base.customer.demographics.credit_rating,
+        count(base.customer.id
+     ) as cnt3,
+    base.customer.demographics.dependent_count,
+            count(base.customer.id
+     ) as cnt4,
+    base.customer.demographics.employed_dependent_count,
+            count(base.customer.id
+     ) as cnt5,
+    base.customer.demographics.college_dependent_count,
+            count(base.customer.id
+     ) as cnt6,
+ORDER BY
+    base.customer.demographics.gender asc,
+    base.customer.demographics.marital_status asc,
+    base.customer.demographics.education_status asc,
+    base.customer.demographics.purchase_estimate asc,
+    base.customer.demographics.credit_rating asc,
+    base.customer.demographics.dependent_count asc,
+    base.customer.demographics.employed_dependent_count asc,
+    base.customer.demographics.college_dependent_count asc,
+
+;
+
