+++ tests/profiling/store_sales.preql
@@ -1,0 +1,44 @@
+import item as item;
+import date as date;
+import time as time;
+import customer as customer;
+import promotion as promotion;
+import customer_demographic as customer_demographic;
+import store as store;
+
+
+key ticket_number int;
+property <ticket_number,item.id>.quantity int;
+property <ticket_number,item.id>.sales_price float;
+property <ticket_number,item.id>.list_price float;
+property <ticket_number,item.id>.ext_sales_price float;
+property <ticket_number,item.id>.ext_wholesale_cost float;
+property <ticket_number,item.id>.ext_list_price float;
+property <ticket_number,item.id>.ext_discount_amount float;
+property <ticket_number,item.id>.coupon_amt float;
+property <ticket_number,item.id>.net_profit float;
+
+auto profit <- ext_list_price - ext_wholesale_cost - ext_discount_amount + ext_sales_price;
+
+datasource store_sales (
+    SS_SOLD_DATE_SK: date.id,
+    SS_SOLD_TIME_SK: time.id,
+    SS_CUSTOMER_SK: ~customer.id,
+    SS_CDEMO_SK: customer_demographic.id,
+    SS_TICKET_NUMBER: ticket_number,
+    SS_ITEM_SK: item.id,
+    SS_SALES_PRICE: sales_price,
+    SS_LIST_PRICE: list_price,
+    SS_EXT_SALES_PRICE: ext_sales_price,
+    SS_EXT_LIST_PRICE: ext_list_price,
+    SS_EXT_WHOLESALE_COST: ext_wholesale_cost,
+    SS_EXT_DISCOUNT_AMOUNT: ext_discount_amount,
+    SS_NET_PROFIT: net_profit,
+    SS_PROMO_SK: promotion.id,
+    SS_QUANTITY: quantity,
+    SS_COUPON_AMT: coupon_amt,
+    SS_STORE_SK: store.id,
+)
+grain (ticket_number, item.id)
+address memory.store_sales
+;
